backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
sqlalchemy
psycopg2-binary
python-dotenv
asyncpg
pydantic
pydantic-settings
openai
python-multipart
email-validator
